refactor(create-house): update import paths for create house form
