+++ src/app/(protected)/dashboard/create/page.tsx
@@ -1,4 +1,4 @@
-import CreateHouseForm from "@/components/house/create-house-form";
+import CreateHouseForm from "@/components/features/group/create-house-form";
 import {
   Card,
   CardContent,
